Support local dev in Caddy via docker-compose
In order to support a local client being swapped in for a stock
client and running in dev mode, we need to run this server on the same
port as the stock client, i.e. 8081. This is because the Caddy
docker-compose setup for auth expects the different services to be at stable
ports. There is a corresponding change in Caddy to support this.